Add typescript support and transpile distribution
