+++ .gitignore
@@ -13,7 +13,7 @@
 /coverage
 
 # production
-/build
+/dist
 
 # misc
 .DS_Store
